# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Keeper Financial API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/candidate/applications": {
      "get": {
        "tags": [
          "candidate applications"
        ],
        "summary": "List Candidate Applications",
        "operationId": "list_candidate_applications_api_v1_candidate_applications_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-dev-auth-sub",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Auth-Sub"
            }
          },
          {
            "name": "x-dev-auth-aal",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "aal1",
              "title": "X-Dev-Auth-Aal"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Candidate access denied"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/candidate/applications/{application_id}/documents": {
      "post": {
        "tags": [
          "candidate documents"
        ],
        "summary": "Upload Candidate Document",
        "operationId": "upload_candidate_document_api_v1_candidate_applications__application_id__documents_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Application Id"
            }
          },
          {
            "name": "x-dev-auth-sub",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Auth-Sub"
            }
          },
          {
            "name": "x-dev-auth-aal",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "aal1",
              "title": "X-Dev-Auth-Aal"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_candidate_document_api_v1_candidate_applications__application_id__documents_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateDocumentResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Candidate ownership or AAL2 denied"
          },
          "404": {
            "description": "Owned application or document not found"
          },
          "409": {
            "description": "Application lifecycle or category limit conflict"
          },
          "422": {
            "description": "Document category, name, type, signature, or size rejected"
          },
          "503": {
            "description": "Private storage or scanning unavailable"
          }
        }
      },
      "get": {
        "tags": [
          "candidate documents"
        ],
        "summary": "List Candidate Documents",
        "operationId": "list_candidate_documents_api_v1_candidate_applications__application_id__documents_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Application Id"
            }
          },
          {
            "name": "x-dev-auth-sub",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Auth-Sub"
            }
          },
          {
            "name": "x-dev-auth-aal",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "aal1",
              "title": "X-Dev-Auth-Aal"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateDocumentList"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Candidate ownership or AAL2 denied"
          },
          "404": {
            "description": "Owned application or document not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/candidate/applications/{application_id}/documents/{document_id}": {
      "delete": {
        "tags": [
          "candidate documents"
        ],
        "summary": "Remove Draft Candidate Document",
        "operationId": "remove_draft_candidate_document_api_v1_candidate_applications__application_id__documents__document_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Application Id"
            }
          },
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          },
          {
            "name": "x-dev-auth-sub",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Auth-Sub"
            }
          },
          {
            "name": "x-dev-auth-aal",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "aal1",
              "title": "X-Dev-Auth-Aal"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Candidate ownership or AAL2 denied"
          },
          "404": {
            "description": "Owned application or document not found"
          },
          "409": {
            "description": "Submitted documents are append-only"
          },
          "503": {
            "description": "Private storage unavailable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApplicationListResponse": {
        "properties": {
          "applications": {
            "items": {
              "$ref": "#/components/schemas/CandidateApplicationResponse"
            },
            "type": "array",
            "title": "Applications"
          }
        },
        "type": "object",
        "required": [
          "applications"
        ],
        "title": "ApplicationListResponse"
      },
      "Body_upload_candidate_document_api_v1_candidate_applications__application_id__documents_post": {
        "properties": {
          "category": {
            "type": "string",
            "title": "Category"
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "category",
          "file"
        ],
        "title": "Body_upload_candidate_document_api_v1_candidate_applications__application_id__documents_post"
      },
      "CandidateApplicationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "recruitment_posting_id": {
            "type": "string",
            "format": "uuid",
            "title": "Recruitment Posting Id"
          },
          "source_posting_slug": {
            "type": "string",
            "title": "Source Posting Slug"
          },
          "source_posting_title": {
            "type": "string",
            "title": "Source Posting Title"
          },
          "source_posting_version": {
            "type": "integer",
            "title": "Source Posting Version"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "state": {
            "type": "string",
            "enum": [
              "draft",
              "submitted",
              "withdrawn"
            ],
            "title": "State"
          },
          "status": {
            "type": "string",
            "enum": [
              "application_started",
              "application_submitted",
              "withdrawn",
              "declined"
            ],
            "title": "Status"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "given_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Given Name"
          },
          "family_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Family Name"
          },
          "preferred_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Name"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "country_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Code"
          },
          "preferred_contact_method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Contact Method"
          },
          "available_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Available From"
          },
          "referral_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Referral Source"
          },
          "referral_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Referral Detail"
          },
          "interest_statement": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interest Statement"
          },
          "relevant_experience": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relevant Experience"
          },
          "employment": {
            "items": {
              "$ref": "#/components/schemas/EmploymentEntryResponse"
            },
            "type": "array",
            "title": "Employment"
          },
          "education": {
            "items": {
              "$ref": "#/components/schemas/EducationEntryResponse"
            },
            "type": "array",
            "title": "Education"
          },
          "privacy_acknowledged": {
            "type": "boolean",
            "title": "Privacy Acknowledged"
          },
          "information_accuracy_confirmed": {
            "type": "boolean",
            "title": "Information Accuracy Confirmed"
          },
          "privacy_disclosure_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Privacy Disclosure Version"
          },
          "privacy_acknowledged_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Privacy Acknowledged At"
          },
          "submitted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          },
          "withdrawn_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Withdrawn At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "recruitment_posting_id",
          "source_posting_slug",
          "source_posting_title",
          "source_posting_version",
          "schema_version",
          "revision",
          "state",
          "status",
          "email",
          "given_name",
          "family_name",
          "preferred_name",
          "phone",
          "city",
          "region",
          "country_code",
          "preferred_contact_method",
          "available_from",
          "referral_source",
          "referral_detail",
          "interest_statement",
          "relevant_experience",
          "employment",
          "education",
          "privacy_acknowledged",
          "information_accuracy_confirmed",
          "privacy_disclosure_version",
          "privacy_acknowledged_at",
          "submitted_at",
          "withdrawn_at",
          "created_at",
          "updated_at"
        ],
        "title": "CandidateApplicationResponse"
      },
      "CandidateDocumentList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CandidateDocumentResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "CandidateDocumentList"
      },
      "CandidateDocumentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "application_id": {
            "type": "string",
            "format": "uuid",
            "title": "Application Id"
          },
          "category": {
            "type": "string",
            "enum": [
              "resume",
              "cover_letter"
            ],
            "title": "Category"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "scan_status": {
            "type": "string",
            "title": "Scan Status"
          },
          "quarantined": {
            "type": "boolean",
            "title": "Quarantined"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "application_id",
          "category",
          "original_filename",
          "content_type",
          "size_bytes",
          "scan_status",
          "quarantined",
          "created_at"
        ],
        "title": "CandidateDocumentResponse"
      },
...
      "EducationEntryResponse": {
        "properties": {
          "institution_name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Institution Name"
          },
          "program_name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Program Name"
          },
          "completion_year": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1900.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Completion Year"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "institution_name",
          "program_name"
        ],
        "title": "EducationEntryResponse"
      },
...
      "EmploymentEntryResponse": {
        "properties": {
          "employer_name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Employer Name"
          },
          "role_title": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Role Title"
          },
          "start_month": {
            "type": "string",
            "pattern": "^(19|20)\\d{2}-(0[1-9]|1[0-2])$",
            "title": "Start Month"
          },
          "currently_employed": {
            "type": "boolean",
            "title": "Currently Employed"
          },
          "end_month": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(19|20)\\d{2}-(0[1-9]|1[0-2])$"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Month"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "employer_name",
          "role_title",
          "start_month",
          "currently_employed"
        ],
        "title": "EmploymentEntryResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}